+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.4.8
 - More complete test pipeline
-- C++ Import
+- C++ & C upgrade to support level Import